requirements.txt
scipy~=1.7.3
pandas~=1.3.5
setuptools~=54.1.1
PyYAML~=5.1